requirements.txt.txt
requests
pysocks
colorama
